feat: add and tested login and logout routes for authentication API endpoint
