+++ backend/src/routes/users.routes.ts
@@ -1,0 +1,17 @@
+import { Router } from 'express';
+import { userController } from '../controllers/user.controller.js';
+import { authMiddleware } from '../middlewares/auth.middleware.js';
+
+const router = Router();
+
+
+// Protected routes
+router.use(authMiddleware);
+
+router.post('/', userController.createUser);
+router.get('/:id', userController.getUserById);
+router.put('/:id', userController.updateUser);
+router.delete('/:id', userController.deleteUser);
+router.get('/me', userController.getMeProfile);
+
+export default router;
